refactor:component configuration change
