+++ src/components/container/groups/GroupListContainer.vue
@@ -5,10 +5,10 @@
 </template>
 
 <script>
-import GroupList from '@/components/presentational/group/GroupList.vue'
-import { groupList } from './demoData.js'
+import GroupList from '@/components/presentational/groups/GroupList.vue'
+import { groupList } from '@/data/demoData.js'
 export default {
-  name: 'MainContainer',
+  name: 'GroupListContainer',
   components: {
     GroupList
   },
