+++ src/components/presentational/groups/GroupList.vue
