+++ src/pages/groups/index.vue
@@ -1,0 +1,13 @@
+<template>
+  <group-list-container />
+</template>
+
+<script>
+import GroupListContainer from '@/components/container/groups/GroupListContainer.vue'
+export default {
+  name: 'GroupsPage',
+  components: {
+    GroupListContainer
+  }
+}
+</script>
