--- src/pages/index.vue
@@ -1,13 +1,0 @@
-<template>
-  <main-container />
-</template>
-
-<script>
-import MainContainer from '@/components/container/MainContainer.vue'
-export default {
-  name: 'MainPage',
-  components: {
-    MainContainer
-  }
-}
-</script>
