+++ src/router/index.ts
@@ -1,14 +1,19 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import Main from '@/pages/index.vue'
+import groups from '@/pages/groups/index.vue'
 
 Vue.use(VueRouter)
 
 const routes = [
   {
     path: '/',
-    name: 'Main',
-    component: Main
+    name: 'main',
+    redirect: '/groups'
+  },
+  {
+    path: '/groups',
+    name: 'groups',
+    component: groups
   },
   {
     path: '*',
